+++ ui/submodules/packages/components/src/components/filters/index.ts
@@ -1,2 +1,1 @@
-export { Filters } from "./filters";
 export type { ComponentsFiltersModel } from "./componentsFiltersModel";