+++ ui/submodules/packages/components/src/scenarios/filters/filtersPart.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo } from "react";
+import React, { useCallback, useMemo } from "react";
 import { flatten, uniq, uniqBy } from "lodash";
 import { useFilterContext } from "../../common";
 import { ScenariosFiltersModel } from "./scenariosFiltersModel";
@@ -17,30 +17,56 @@ export function FiltersPart({ isLoading, data = [] }: { data: RowType[]; isLoadi
     const { t } = useTranslation();
     const { data: userData } = useUserQuery();
 
-    const filterableKeys = useMemo(() => ["createdBy"], []);
+    const filterableKeys = useMemo(() => ["createdBy", "modifiedBy"], []);
     const filterableValues = useMemo(() => {
         const entries = filterableKeys.map((k) => [k, uniq(flatten(data.map((v) => v[k]))).sort()]);
-        return Object.fromEntries([
-            ...entries,
-            [
-                "status",
-                uniqBy(
-                    data.map((v) => ({ name: v.state?.status.name, icon: v.state?.icon })),
-                    "name",
-                ).sort(),
-            ],
-            ["processCategory", (userData?.categories || []).map((name) => ({ name }))],
-        ]);
+        return {
+            ...Object.fromEntries(entries),
+            author: uniq(["modifiedBy", "createdBy"].flatMap((k) => data.flatMap((v) => v[k])))
+                .sort()
+                .map((v) => ({ name: v })),
+            status: uniqBy(
+                data.map((v) => ({ name: v.state?.status.name, icon: v.state?.icon })),
+                "name",
+            ).sort(),
+            processCategory: (userData?.categories || []).map((name) => ({ name })),
+        };
     }, [data, filterableKeys, userData?.categories]);
 
     const statusFilters: Array<keyof ScenariosFiltersModel> = ["HIDE_DEPLOYED", "HIDE_NOT_DEPLOYED"];
-    const { getFilter, setFilter } = useFilterContext<ScenariosFiltersModel>();
+    const { getFilter, setFilter, activeKeys } = useFilterContext<ScenariosFiltersModel>();
 
     const otherFilters: Array<keyof ScenariosFiltersModel> = ["HIDE_SCENARIOS", "HIDE_FRAGMENTS", "HIDE_ACTIVE", "SHOW_ARCHIVED"];
 
+    const getLabel = useCallback(
+        (name: keyof ScenariosFiltersModel, value?: string | number) => {
+            switch (name) {
+                case "HIDE_ACTIVE":
+                    return t("table.filter.desc.HIDE_ACTIVE", "Active hidden");
+                case "HIDE_FRAGMENTS":
+                    return t("table.filter.desc.HIDE_FRAGMENTS", "Fragments hidden");
+                case "HIDE_SCENARIOS":
+                    return t("table.filter.desc.HIDE_SCENARIOS", "Scenarios hidden");
+                case "SHOW_ARCHIVED":
+                    return t("table.filter.desc.SHOW_ARCHIVED", "Archived visible");
+                case "HIDE_DEPLOYED":
+                    return t("table.filter.desc.HIDE_DEPLOYED", "Deployed hidden");
+                case "HIDE_NOT_DEPLOYED":
+                    return t("table.filter.desc.HIDE_NOT_DEPLOYED", "Not deployed hidden");
+            }
+
+            if (value?.toString().length) {
+                return value;
+            }
+
+            return name;
+        },
+        [t],
+    );
+
     return (
         <>
-            <QuickFilter isLoading={isLoading}>
+            <QuickFilter<ScenariosFiltersModel> isLoading={isLoading} filter="NAME">
                 <Stack direction="row" spacing={1} p={1} alignItems="center" divider={<Divider orientation="vertical" flexItem />}>
                     {/*<FilterMenu label={t("table.filter.STATUS", "Status")} count={getFilter("STATUS", true).length}>*/}
                     {/*    <SimpleOptionsStack*/}
@@ -56,7 +82,7 @@ export function FiltersPart({ isLoading, data = [] }: { data: RowType[]; isLoadi
                     <FilterMenu label={t("table.filter.CATEGORY", "Category")} count={getFilter("CATEGORY", true).length}>
                         <SimpleOptionsStack
                             label={t("table.filter.CATEGORY", "Category")}
-                            options={filterableValues["processCategory"]}
+                            options={filterableValues.processCategory}
                             value={getFilter("CATEGORY", true)}
                             onChange={setFilter("CATEGORY")}
                         />
@@ -64,7 +90,7 @@ export function FiltersPart({ isLoading, data = [] }: { data: RowType[]; isLoadi
                     <FilterMenu label={t("table.filter.CREATED_BY", "Author")} count={getFilter("CREATED_BY", true).length}>
                         <SimpleOptionsStack
                             label={t("table.filter.CREATED_BY", "Author")}
-                            options={filterableValues["createdBy"].map((name) => ({ name }))}
+                            options={filterableValues.author}
                             value={getFilter("CREATED_BY", true)}
                             onChange={setFilter("CREATED_BY")}
                         />
@@ -82,7 +108,7 @@ export function FiltersPart({ isLoading, data = [] }: { data: RowType[]; isLoadi
                     </FilterMenu>
                 </Stack>
             </QuickFilter>
-            <ActiveFilters />
+            <ActiveFilters getLabel={getLabel} activeKeys={activeKeys.filter((k) => k !== "NAME" && k !== "SORT_BY")} />
         </>
     );
 }