+++ utils/flink-components-testkit/src/main/scala/pl/touk/nussknacker/engine/flink/util/test/FlinkTestScenarioRunner.scala
@@ -18,7 +18,7 @@ import pl.touk.nussknacker.engine.process.helpers.SinkForType
 import pl.touk.nussknacker.engine.process.registrar.FlinkProcessRegistrar
 import pl.touk.nussknacker.engine.testing.LocalModelData
 import pl.touk.nussknacker.engine.testmode.{TestComponentHolder, TestComponentsHolder}
-import pl.touk.nussknacker.engine.util.test.TestScenarioRunner
+import pl.touk.nussknacker.engine.util.test.ClassBaseTestScenarioRunner
 
 import scala.reflect.ClassTag
 
@@ -32,13 +32,13 @@ object testComponents {
 
 case class SinkForList[T]() extends SinkForType[List[T]]
 
-class FlinkTestScenarioRunner(val components: List[ComponentDefinition], val config: Config, flinkMiniCluster: FlinkMiniClusterHolder) extends TestScenarioRunner {
+class FlinkTestScenarioRunner(val components: List[ComponentDefinition], val config: Config, flinkMiniCluster: FlinkMiniClusterHolder) extends ClassBaseTestScenarioRunner {
 
   var testComponentHolder: TestComponentHolder = _
 
-  override def runWithData[T: ClassTag, Result](scenario: EspProcess, data: List[T]): List[Result] = {
+  override def runWithData[I:ClassTag, R](scenario: EspProcess, data: List[I]): List[R] = {
 
-    implicit val typeInf: TypeInformation[T] = TypeInformation.of(implicitly[ClassTag[T]].runtimeClass.asInstanceOf[Class[T]])
+    implicit val typeInf: TypeInformation[I] = TypeInformation.of(implicitly[ClassTag[I]].runtimeClass.asInstanceOf[Class[I]])
     val modelData = LocalModelData(config, new EmptyProcessConfigCreator)
     val componentsWithData = testComponents.withDataList(data)
     val componentsWithTestComponents = this.components ++ componentsWithData
@@ -49,7 +49,7 @@ class FlinkTestScenarioRunner(val components: List[ComponentDefinition], val con
     val registrar = FlinkProcessRegistrar(new FlinkProcessCompilerWithTestComponents(testComponentHolder, modelData), ExecutionConfigPreparer.unOptimizedChain(modelData))
     registrar.register(new StreamExecutionEnvironment(env), scenario, ProcessVersion.empty, DeploymentData.empty, Some(testComponentHolder.runId))
     env.executeAndWaitForFinished(scenario.id)()
-    testComponentHolder.results(testComponentHolder.runId).map((k: Any) => k.asInstanceOf[Result])
+    testComponentHolder.results(testComponentHolder.runId).map((k: Any) => k.asInstanceOf[R])
   }
 
 }